[UPDATE] lotto step3 구현 * LottoGame - LottoGenerator를 생성자통해 의존성 주입하여 게임 진행하는 동안 사용 * LottoGenerator - 추상클래스로 변경, generateLotto 추상메소드 - FixedLottoGenerator, RandomLottoGenerator가 상속 * Lotto 클래스의 로또번호를 담은 자료구조 Set으로 변경 및 중복 확인 로직 제거 * PurchasedLottos - 수동구매 로또개수 인스턴스 변수추가 - 수동구매 관련 로직 추가 * Money - 금액을 의미하는 원시값 래핑 * LottosResult - 총수익 계산 로직 스트림 사용하는 것으로 변경 - 수익률 계산하는 로직 부분적으로 Money에 위임

+++ src/main/java/lotto/Prize.java
@@ -9,10 +9,10 @@ public enum Prize {
     NO_PRIZE(0, 0, false);
 
     private final int numberOfMatch;
-    private final long prizeMoney;
+    private final int prizeMoney;
     private final boolean isMatchingBonus;
 
-    Prize(int numberOfMatch, long prizeMoney, boolean isMatchingBonus) {
+    Prize(int numberOfMatch, int prizeMoney, boolean isMatchingBonus) {
         this.numberOfMatch = numberOfMatch;
         this.prizeMoney = prizeMoney;
         this.isMatchingBonus = isMatchingBonus;
@@ -28,13 +28,13 @@ public enum Prize {
     }
 
     private boolean isEqual(int numberOfMatch, boolean isMatchingBonus) {
-        if(numberOfMatch == Prize.SECOND_PRIZE.numberOfMatch) {
+        if (numberOfMatch == Prize.SECOND_PRIZE.numberOfMatch) {
             return this.numberOfMatch == numberOfMatch && this.isMatchingBonus == isMatchingBonus;
         }
         return this.numberOfMatch == numberOfMatch;
     }
 
-    public long calculatePrizeMoneyBy(int count) {
+    public int calculatePrizeMoneyBy(int count) {
         return count * prizeMoney;
     }
